+++ src/PagesInterface.php
@@ -1,20 +1,11 @@
 <?php
 
-declare(strict_types = 1);
-
 namespace iio\libmergepdf;
 
-interface PagesInterface extends \IteratorAggregate
+interface PagesInterface
 {
     /**
-     * Get iterator of page numbers
-     *
-     * @return iterable & int[]
+     * @return int[]
      */
-    public function getIterator(): iterable;
-
-    /**
-     * Check if this collection is empty
-     */
-    public function isEmpty(): bool;
+    public function getPageNumbers(): array;
 }